requirements.txt
# Web工具依赖
requests>=2.28.0
beautifulsoup4>=4.11.0

# 图像处理依赖
Pillow>=9.0.0

# 数据处理依赖
pandas>=1.5.0
pyyaml>=6.0
python-dateutil>=2.8.0

# LLM 依赖
openai>=1.0.0

# 工具依赖
numpy>=1.23.0
pytz>=2022.1

# 开发依赖
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0 